Add original query str to the queryApi object
In order to gain access to the original query string
inside the resultBrowsers, we need to preserve the query
string inside the queryApi object.  This is not ideal,
because the api object is both the API wrapper, with
the information HOW to get the data, as well as the
storage for the RESULT data itself.  I think the API
object should return an object with the request,
server's response, and any additional metrics,
thus allowing everyone to share in on it.  But such
change should probably happen after we migrate to Babel,
and start using proper Promises.

Change-Id: I4da3dd36ccec5ae5d91e210d8f344b8228008f9b